Cleaned bugs, overhauled UI, apply themes for issue#41

+++ app/src/main/java/com/csce5430sec7proj/pethelper/ui/services/ServiceDetailsScreen.kt
@@ -18,16 +18,17 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.unit.sp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.compose.runtime.collectAsState
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.size
+import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.layout.ContentScale
@@ -37,6 +38,10 @@ import androidx.compose.ui.unit.dp
 import com.csce5430sec7proj.pethelper.R
 import com.csce5430sec7proj.pethelper.data.entities.Service
 import com.csce5430sec7proj.pethelper.data.entities.ServiceCategory
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.graphics.Color
+import com.csce5430sec7proj.pethelper.ui.components.LabelWithText
 
 
 @Composable
@@ -49,14 +54,22 @@ fun ServiceDetailsScreen(
     val viewModel: ServicesViewModel = viewModel()
     val serviceState = viewModel.state.collectAsState().value
     val service: Service? = serviceState.services.find { it.id == serviceId }
+    // State variables to handle confirmations
+    var showDeleteConfirmation by remember { mutableStateOf(false) }
 
     Scaffold(
         modifier = modifier.fillMaxSize(),
         floatingActionButton = {
-            FloatingActionButton(onClick = {
-                service?.let { onNavigate(it.id) }
-            }) {
-                Icon(imageVector = Icons.Default.Edit, contentDescription = stringResource(id = R.string.Edit))
+            FloatingActionButton(
+                onClick = { service?.let { onNavigate(it.id) } },
+                containerColor = MaterialTheme.colorScheme.secondaryContainer,
+                contentColor = MaterialTheme.colorScheme.onSecondaryContainer
+            ) {
+                Icon(
+                    imageVector = Icons.Default.Edit,
+                    contentDescription = stringResource(id = R.string.edit_service),
+                    tint = MaterialTheme.colorScheme.onPrimary
+                )
             }
         }
     ) { padding ->
@@ -72,10 +85,10 @@ fun ServiceDetailsScreen(
                 modifier = Modifier
                     .size(50.dp)
                     .clip(CircleShape)
-                    .background(Color.Gray)
+                    .background(MaterialTheme.colorScheme.surface)
                     .padding(padding)
             ) {
-                val iconResId = when (service?.category)  {
+                val iconResId = when (service?.category) {
                     ServiceCategory.GROOMER -> R.drawable.ic_bird
                     ServiceCategory.TRAINER -> R.drawable.ic_cat
                     ServiceCategory.DAYCARE -> R.drawable.ic_cow
@@ -93,81 +106,75 @@ fun ServiceDetailsScreen(
                     contentScale = ContentScale.Inside
                 )
             }
+
             Spacer(modifier = Modifier.height(16.dp))
+
             service?.let {
                 // Service Name
                 Text(
                     text = it.name,
-                    style = MaterialTheme.typography.headlineLarge.copy(fontSize = 32.sp),
-                    color = Color.Black,
+                    style = MaterialTheme.typography.headlineLarge,
+                    color = MaterialTheme.colorScheme.onPrimaryContainer,
                     modifier = Modifier.padding(top = 16.dp)
                 )
             }
+
             Spacer(modifier = Modifier.height(16.dp))
+
             // Align remaining content to the left
             Column(
                 modifier = Modifier.fillMaxWidth(),
                 horizontalAlignment = Alignment.Start
             ) {
                 service?.let {
-                // Service Description
-                Text( text = stringResource(id = R.string.description_hint) )
-                Text(
-                    text = service.description ?: "",
-                    style = MaterialTheme.typography.headlineLarge.copy(fontSize = 22.sp),
-                    color = Color.Black,
-                    modifier = Modifier.padding(top = 16.dp)
-                )
-                Spacer(modifier = Modifier.height(16.dp))
-                // Service Address
-                Text( text = stringResource(id = R.string.address_hint) )
-                Text(
-                    text = service.address ?: "",
-                    style = MaterialTheme.typography.headlineLarge.copy(fontSize = 22.sp),
-                    color = Color.Black,
-                    modifier = Modifier.padding(top = 16.dp)
-                )
-                Spacer(modifier = Modifier.height(16.dp))
-                // Service Phone
-                Text( text = stringResource(id = R.string.phone_hint) )
-                Text(
-                    text = service.phone ?: "",
-                    style = MaterialTheme.typography.bodyLarge.copy(fontSize = 22.sp),
-                    color = Color.Black
-                )
-                Spacer(modifier = Modifier.height(16.dp))
-                // Service Secondary Phone
-                Text( text = stringResource(id = R.string.secondary_phone_hint) )
-                Text(
-                    text = service.secondaryPhone ?: "",
-                    style = MaterialTheme.typography.bodyLarge.copy(fontSize = 22.sp),
-                    color = Color.Black
-                )
-                Spacer(modifier = Modifier.height(16.dp))
-                // Service Email
-                Text( text = stringResource(id = R.string.email_hint) )
-                Text(
-                    text = service.email ?: "",
-                    style = MaterialTheme.typography.bodyLarge.copy(fontSize = 22.sp),
-                    color = Color.Black
-                )
-                Spacer(modifier = Modifier.height(16.dp))
-                // Service Notes
-                Text( text = stringResource(id = R.string.notes_hint) )
-                Text(
-                    text = service.notes ?: "",
-                    style = MaterialTheme.typography.bodyLarge.copy(fontSize = 22.sp),
-                    color = Color.Black
-                )
-            } ?: run {
-                Text(
-                    text = stringResource(id = R.string.loading_service_details),
-                    style = MaterialTheme.typography.bodyLarge.copy(fontSize = 20.sp),
-                    color = Color.Gray,
-                    modifier = Modifier.padding(top = 16.dp)
-                )
+                    // Service Description
+                    LabelWithText(
+                        label = stringResource(id = R.string.description_hint),
+                        text = it.description ?: ""
+                    )
+
+                    // Service Address
+                    LabelWithText(
+                        label = stringResource(id = R.string.address_hint),
+                        text = it.address ?: ""
+                    )
+
+                    // Service Phone
+                    LabelWithText(
+                        label = stringResource(id = R.string.phone_hint),
+                        text = it.phone ?: ""
+                    )
+
+                    // Service Secondary Phone
+                    LabelWithText(
+                        label = stringResource(id = R.string.secondary_phone_hint),
+                        text = it.secondaryPhone ?: ""
+                    )
+
+                    // Service Email
+                    LabelWithText(
+                        label = stringResource(id = R.string.email_hint),
+                        text = it.email ?: ""
+                    )
+
+                    // Service Notes
+                    LabelWithText(
+                        label = stringResource(id = R.string.notes_hint),
+                        text = it.notes ?: ""
+                    )
+                } ?: run {
+                    Text(
+                        text = stringResource(id = R.string.loading_service_details),
+                        style = MaterialTheme.typography.bodyLarge,
+                        color = MaterialTheme.colorScheme.onBackground,
+                        modifier = Modifier.padding(top = 16.dp)
+                    )
+                }
             }
+
             Spacer(modifier = Modifier.height(25.dp))
+
+            // Action Buttons
             Row(
                 modifier = Modifier
                     .fillMaxWidth()
@@ -175,19 +182,61 @@ fun ServiceDetailsScreen(
                 horizontalArrangement = Arrangement.SpaceEvenly,
                 verticalAlignment = Alignment.CenterVertically
             ) {
+                // Delete a service
                 Button(
                     onClick = {
                         service?.let {
-                            viewModel.deleteService(it)
-                            onNavigateBack()
+                            // Show the confirmation dialog
+                            showDeleteConfirmation = true
                         }
                     },
-                    modifier = Modifier.weight(1f).padding(start = 8.dp),
-                    colors = ButtonDefaults.buttonColors(containerColor = Color.Red)
+                    colors = ButtonDefaults.buttonColors(
+                        containerColor = MaterialTheme.colorScheme.error,
+                        contentColor = MaterialTheme.colorScheme.onError
+                    )
                 ) {
                     Text(text = stringResource(id = R.string.delete))
                 }
-            }
+
+                // Delete Confirmation Dialog
+                if (showDeleteConfirmation) {
+                    AlertDialog(
+                        onDismissRequest = { showDeleteConfirmation = false },
+                        title = {
+                            Text(
+                                text = stringResource(id = R.string.confirm_delete_title),
+                                style = MaterialTheme.typography.titleLarge,
+                              //  color = MaterialTheme.colorScheme.onSurface
+                            )
+                        },
+                        text = {
+                            Text(
+                                text = stringResource(id = R.string.confirm_delete_message_service),
+                                style = MaterialTheme.typography.bodyLarge,
+                              //  color = MaterialTheme.colorScheme.onSurface
+                            )
+                        },
+                        confirmButton = {
+                            Button(
+                                onClick = {
+                                    service?.let {
+                                        viewModel.deleteService(it)
+                                        onNavigateBack()
+                                    }
+                                    showDeleteConfirmation = false
+                                },
+                                colors = ButtonDefaults.buttonColors(containerColor = Color.Red)
+                            ) {
+                                Text(text = stringResource(id = R.string.delete))
+                            }
+                        },
+                        dismissButton = {
+                            Button(onClick = { showDeleteConfirmation = false }) {
+                                Text(text = stringResource(id = R.string.cancel))
+                            }
+                        }
+                    )
+                }
             }
         }
     }
